% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data1}
\alias{example_data1}
\title{A numeric vector produced by `rnorm`}
\format{
A numeric vector, length 25
\describe{
  A numeric vector generated by `rnorm(25)`
}
}
\usage{
example_data1
}
\description{
A numeric vector produced by `rnorm`
}
\keyword{datasets}
