% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_redo_state.R
\name{get.redo.state}
\alias{get.redo.state}
\title{Get state of Redo Variable(s)}
\usage{
get.redo.state(redo.vars)
}
\arguments{
\item{redo.vars}{A character vector of the variable names you want to get the state of.}
}
\description{
Get the state (TRUE or FALSE) of one or more redo variables.
}
\examples{
chunk.names <- c("chunk1", "chunk2")

assign.redo(chunk.names)

get.redo.sate("chunk1")
}
