% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_redos.R
\name{list.redos}
\alias{list.redos}
\title{List Current Redo Variables}
\usage{
list.redos()
}
\description{
Lists the current variables in the `redo` environment.
}
\examples{
chunk.names <- c("chunk1", "chunk2")

assign.redo(chunk.names)

list.redos()
}
