% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redo_env.R
\docType{data}
\name{redo}
\alias{redo}
\title{Redo Environment}
\format{
An object of class \code{environment} of length 0.
}
\usage{
redo
}
\description{
Create environment for tracking various redo variables (set to TRUE or FALSE)
}
\keyword{datasets}
