% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_redo.R
\name{remove.redo}
\alias{remove.redo}
\title{Remove Redo Variable(s)}
\usage{
remove.redo(redo.vars)
}
\arguments{
\item{redo.vars}{A character vector of the variable names you want to remove.}
}
\description{
Remove one or more redo variables from the redo environment.
}
\examples{
chunk.names <- c("chunk1", "chunk2")

assign.redo(chunk.names)
list.redos()

remove.redo("chunk1")
list.redos()
}
