% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_redo_true_false.R
\name{set.redo.true}
\alias{set.redo.true}
\title{Set Redo Variables to TRUE or FALSE}
\usage{
set.redo.true(redo.vars)
}
\arguments{
\item{redo.vars}{A character vector of the variable names you want to give TRUE or FALSE to}
}
\description{
Set the state (TRUE or FALSE) of one or more redo variables. This can also be acheived with assign_redo(), but these functions require only one argument.
}
\examples{
chunk.names <- c("chunk1", "chunk2")

assign.redo(chunk.names)

set.redo.true("chunk1")
}
