% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redo_if.R
\name{redo.if}
\alias{redo.if}
\title{Run code if redo variable is TRUE}
\usage{
redo.if(redo.var, save.file, code, force = FALSE)
}
\arguments{
\item{redo.var}{The redo variable name you want to reference.}

\item{save.file}{The RDS file to which you want to save the results or from which you want to read prior results. If `.rds` is not at the end of the file, it will be appended.}

\item{code}{The code you want to execute within the statement which should produce the object you want to return and save when the redo variables is TRUE.}
}
\description{
Basically executes an IF/ELSE statement to either re-run code (if the redo variable is TRUE) or return data saved in an RDS file (if the redo variable is FALSE).
}
\examples{
chunk.names <- c("chunk1", "chunk2")

assign.redo(chunk.names)

toggle.redo("chunk1")

if (redo$chunk1) {
  x <- rnorm(100)
} else {
  x <- data(example_data1)
}

x <- redo.if("chunk1", save.file = "test_file.rds", {rnorm(50)})
}
